bytter på navn og refaktoriserer litt på informasjonspanel (den nye uten Collapse)
