--- web/src/frontend/src/nav-soknad/components/informasjonspanelTo/informasjonspanelTo.less
@@ -1,43 +1,0 @@
-.informasjonspanelTo {
-  border-radius: 8px;
-  border: solid 2px #ffd399;
-  min-height: 80px;
-  display: flex;
-  align-items: center;
-  margin-left: 2rem;
-  margin-right: 2rem;
-  padding-top: 2rem;
-  padding-bottom: 2rem;
-}
-
-.informasjonspanelTo__ikon {
-  height: 80px;
-  transform: translate(-40px);
-}
-
-.informasjonspanelTo__tekst {
-  transform: translate(-23px);
-  padding-right: 20px;
-}
-
-.informasjonspanelToWrapper {
-  padding-bottom: 5rem;
-}
-
-
-//
-//&>div {
-//  position: absolute;
-//  left: 0;
-//  top: 50%;
-//  transform: translate(-50%, -50%);
-//
-//  &>img {
-//    position: absolute;
-//    left: 50%;
-//    top: 50%;
-//    transform: translate(-50%, -40%);
-//    height: 4.5rem;
-//  }
-//
-//}
