+++ web/src/frontend/src/nav-soknad/styles/nav-soknad.less
@@ -5,7 +5,7 @@
 @import "../components/dialog/dialog.less";
 @import "../components/feilside/feilside.less";
 @import "../components/infoblokk/infoblokk.less";
-@import "../components/informasjonspanel/informasjonspanel.less";
+@import "../components/informasjonspanelEkspanderbart/informasjonspanelEkspanderbart.less";
 @import "../components/knapperad/knapperad.less";
 @import "../components/labelMedHjelpetekst/labelMedHjelpetekst.less";
 @import "../components/lenkeknapp/lenkeknapp.less";
@@ -24,7 +24,7 @@
 @import "../components/svg/brevkonvolutt.less";
 @import "../components/svg/ella.less";
 @import "../components/fadecollapse/fadeCollapse.less";
-@import "../components/informasjonspanelTo/informasjonspanelTo.less";
+@import "../components/informasjonspanel/informasjonspanel.less";
 
 
 @import "../components/animasjoner/ellablunk/ellablunk.less";
